+++ src/module/Accounting/config/module.config.php
@@ -2,24 +2,24 @@
 return array(
 	'router' => array(
 		'routes' => array(
-            'accounting-home' => array(
-                'type' => 'Zend\Mvc\Router\Http\Literal',
-                'options' => array(
-                    'route'    => '/accounting/',
-                    'defaults' => array(
-                        'controller' => 'Accounting\Controller\Index',
-                        'action'     => 'index',
-                    ),
-                ),
-            ),
-			'accounts' => array (
-				'type'    => 'segment',
-				'options' => array (
-					'route'       => '/accounting/accounts[/:id][/:controller]',
-					'constraints' => array (
-						'id'     => '[0-9a-z\-]+',
+			'accounting-home' => array(
+				'type' => 'Zend\Mvc\Router\Http\Literal',
+				'options' => array(
+					'route'	   => '/accounting/',
+					'defaults' => array(
+						'controller' => 'Accounting\Controller\Index',
+						'action'	 => 'index',
 					),
-					'defaults'    => array (
+				),
+			),
+			'accounts' => array (
+				'type'	  => 'segment',
+				'options' => array (
+					'route'		  => '/accounting/accounts[/:id][/:controller]',
+					'constraints' => array (
+						'id'	 => '[0-9a-z\-]+',
+					),
+					'defaults'	  => array (
 						'__NAMESPACE__' => 'Accounting\Controller',
 						'controller' => 'Accounts'
 					),
@@ -77,7 +77,4 @@ return array(
 		'Accounting\CreatePersonalAccountListener',
 		'Accounting\CreateOrganizationAccountListener'
 	)
-	
-	
-	
 );
