+++ src/module/Accounting/src/Accounting/Account.php
@@ -1,5 +1,5 @@
 <?php
-namespace Ora\Accounting;
+namespace Accounting;
 
 use Rhumsaa\Uuid\Uuid;
 use Ora\DomainEntity;
