+++ src/module/Accounting/src/Accounting/Balance.php
@@ -1,5 +1,5 @@
 <?php
-namespace Ora\Accounting;
+namespace Accounting;
 
 /**
  *
