+++ src/module/Accounting/src/Accounting/Controller/StatementsController.php
@@ -4,7 +4,7 @@ namespace Accounting\Controller;
 use Application\Controller\AuthController;
 
 use ZendExtension\Mvc\Controller\AbstractHATEOASRestfulController;
-use Ora\Accounting\AccountService;
+use Accounting\Service\AccountService;
 use Accounting\View\StatementJsonModel;
 use BjyAuthorize\Service\Authorize;
