+++ src/module/Accounting/src/Accounting/OutgoingCreditsTransferred.php
@@ -1,5 +1,5 @@
 <?php
-namespace Ora\Accounting;
+namespace Accounting;
 
 use Prooph\EventSourcing\AggregateChanged;
